% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_packages.R
\name{all_packages}
\alias{all_packages}
\title{Details packages available from <opendata.nhs.scot>.}
\usage{
all_packages(contains = NULL, limit = 1000L)
}
\arguments{
\item{contains}{a character string containing an expression to be used as
search criteria against the packages 'title' field.}

\item{limit}{a numeric value specifying the maximum number of rows to be
returned. Defaults to 1000L.}
}
\value{
a data.frame containing the names of all available packages and their
package ids, or those whose name contains the string specified in the
\code{contains} argument.
}
\description{
Details all packages available from <opendata.nhs.scot> in a
data.frame along with the package id, with the option to limit results based
on a search term.
}
\examples{
\dontrun{
all_packages()
all_packages(contains = "standard-populations")
}

}
