% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_resources.R
\name{all_resources}
\alias{all_resources}
\title{Provides an overview of all 'resources' available from opendata.nhs.scot}
\usage{
all_resources(package_contains = NULL, resource_contains = NULL)
}
\arguments{
\item{package_contains}{character string treated as a regular expression to
be matched against the package names returned.}

\item{resource_contains}{character string treated as a regular expression to
be matched against the resource names returned.}
}
\value{
a table containing the names of all available packages and resources,
or those containing the string specified in the contains arguments, along
with the resource ID.
}
\description{
Provides an overview of all 'resources' available from
opendata.nhs.scot. With the option to limit results based on both package
and resource names.
}
\examples{
\dontrun{
all_resources()
all_resources(package_contains = "population")
all_resources(package_contains = "population", resource_contains = "GP")
}
}
