% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap_url.R
\name{cap_url}
\alias{cap_url}
\title{Produces error if input exceeds 2000 characters.}
\usage{
cap_url(x)
}
\arguments{
\item{x}{a character string to check.}
}
\value{
invisible.
}
\description{
Used to ensure constructed URLs do not exceed 2000 characters.
}
