% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_items.R
\name{data_items}
\alias{data_items}
\title{Get a table of data items and their types for a given resource.}
\usage{
data_items(resource)
}
\arguments{
\item{resource}{a valid resource id.}
}
\value{
a table containing the names and types of all data items available
for the chosen resource.
}
\description{
Get a table of data items and their types for a given resource.
}
\examples{
\dontrun{
data_items(resource="8f7b64b1-eb53-43e9-b888-45af0bc25505")
}
}
