% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_error.R
\name{detect_error}
\alias{detect_error}
\title{Detects http errors and provides enhanced details.}
\usage{
detect_error(result)
}
\arguments{
\item{result}{A http response.}
}
\value{
invisible
}
\description{
Detects http errors and provides enhanced details.
}
