% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets.R
\name{get_dataset}
\alias{get_dataset}
\title{Get selected resources or all resources within a package.}
\usage{
get_dataset(package, resource = NULL, limit = Inf)
}
\arguments{
\item{package}{A resource id identifying the package for which all resources
should be returned or the id of the package containing te resource/s passed
in the resource argument.}

\item{resource}{Optionally specify resource id/s identifying the data set/s
to be returned.}

\item{limit}{Specify the maximum number of records to be returned.}
}
\value{
A list containing all the resources within a package, or those
specified.
}
\description{
Get data from one or more resources, or all resources within a
package, as a list with each resource in tabular format.
}
