% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrow_resource.R
\name{nrow_resource}
\alias{nrow_resource}
\title{Get the number of rows present in a resource.}
\usage{
nrow_resource(resource)
}
\arguments{
\item{resource}{A character string containing the resource id of the data set
to be returned.}
}
\value{
An integer of length 1 indicating the number of rows present in the
specified resource.
}
\description{
Get the number of rows present in a resource.
}
\examples{
\dontrun{
nrow_resource(resource = "edee9731-daf7-4e0d-b525-e4c1469b8f69")
}
}
