% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_metadata.R
\name{package_metadata}
\alias{package_metadata}
\title{Get metadata for a package.}
\usage{
package_metadata(package)
}
\arguments{
\item{package}{A character vector of length 1 specifying a package id or name
which identifies the package for which metadata should be returned.}
}
\value{
a list containing the package metadata.
}
\description{
Get a specified packages metadata as a list.
}
\examples{
\dontrun{
package_metadata(package = "standard-populations")
package_metadata(package = "edee9731-daf7-4e0d-b525-e4c1469b8f69")
}

}
