% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_where.R
\name{parse_where}
\alias{parse_where}
\title{Attempt to parse the 'WHERE' element of a simple SQL SELECT style query to a
format compatible with the CKAN API.}
\usage{
parse_where(x, fields)
}
\arguments{
\item{x}{a string containing the 'WHERE' element of a SQL SELECT query.}

\item{fields}{character vector of fields confirmed to be present in the
target resource. Typically obtained with \code{field_metadata()}.}
}
\value{
a string.
}
\description{
Attempt to parse the 'WHERE' element of a simple SQL SELECT style query to a
format compatible with the CKAN API.
}
