% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_queries.R
\name{prep_nosql_query}
\alias{prep_nosql_query}
\title{Prepare an API query without SQL.}
\usage{
prep_nosql_query(resource, fields, limit)
}
\arguments{
\item{resource}{A resource id identifying the data set to be returned.}

\item{fields}{Names of fields to be included in the returned table.}

\item{limit}{Specify the maximum number of records to be returned.}
}
\value{
A string containing the prepared query.
}
\description{
Prepare an API query without SQL.
}
