% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_queries.R
\name{prep_sql_query}
\alias{prep_sql_query}
\title{Prepare an API query with SQL.}
\usage{
prep_sql_query(resource, fields, limit, offset, where)
}
\arguments{
\item{resource}{a character string specifying resource id of the data set
to be returned.}

\item{fields}{a character vector specifying the names of fields to be
included in the returned data.}

\item{limit}{A numeric value specifying the maximum number of rows to be
returned.}

\item{offset}{A numeric value specifying the number of rows to skip.}

\item{where}{A character string containing the 'WHERE' element of a simple
SQL SELECT style query. Field names must be double quoted (\code{"}), non
numeric values must be single quoted (\code{"}), and both single and double
quotes must be delimited. Example; \code{where = "\"AgeGroup\" = 
 \'45-49 years\\'"}.}
}
\value{
A character string containing the prepared query.
}
\description{
Prepare an API query with SQL.
}
