% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_metadata.R
\name{resource_metadata}
\alias{resource_metadata}
\title{Get a table of available fields and their field_metadata for a given
resource.}
\usage{
resource_metadata(resource)
}
\arguments{
\item{resource}{a valid resource id.}
}
\value{
a table containing the names and types of all data items available
for the chosen resource.
}
\description{
Get a table of available fields and their field_metadata for a given
resource.
}
\examples{
\dontrun{
field_metadata(resource="8f7b64b1-eb53-43e9-b888-45af0bc25505")
}

}
