% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_id.R
\name{valid_id}
\alias{valid_id}
\title{Basic check of whether the characters in a string is 36}
\usage{
valid_id(x)
}
\arguments{
\item{x}{a string}
}
\value{
logical value indicating whether the srting checked consists of 36
characters.
}
\description{
Basic check of whether the characters in a string is 36
}
