% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_packages.R
\name{all_packages}
\alias{all_packages}
\title{Lists 'packages' available from opendata.nhs.scot}
\usage{
all_packages(contains = NULL)
}
\arguments{
\item{contains}{character string treated as a regular expression to be
matched against the package names returned.}
}
\value{
a character vector containing the names of all available packages,
or those containing the string specified in the contains argument.
}
\description{
Lists all 'packages' available from opendata.nhs.scot. With the
option to limit results based on a regular expression.
}
\examples{
\dontrun{
all_packages()
all_packages(contains = "population")
}

}
