% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get data from a resource.}
\usage{
get_data(resource, fields = NULL, limit = NULL, where = NULL)
}
\arguments{
\item{resource}{A resource id identifying the data set to be returned.}

\item{fields}{Names of fields to be included in the returned table.}

\item{limit}{Specify the maximum number of records to be returned, default
NULL value returns all records. Note that specifying a limit > 99,999 or no
limit forces the use of SQL, which carries a small performance penalty.}

\item{where}{A string containing the 'WHERE' element of a simple SQL SELECT
style query. Field names must be double quoted, non numeric values must be
single quoted, and single and double quotes must be delimited.}
}
\value{
A data.frame containing the selected columns from the specified data
set.
}
\description{
Get data from a resource in tabular format.
}
\examples{
\dontrun{
get_data(
  resource = "42f17a3c-a4db-4965-ba68-3dffe6bca13a",
  fields = c("Product", "Dose", "CumulativeNumberVaccinated"),
  limit = 10,
  where = "\"Product\" = \'Total\'"
  )
}

}
