% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get data from a resource.}
\usage{
get_data(resource, data_items = NULL, limit = NULL, default_ord = FALSE)
}
\arguments{
\item{resource}{A resource id identifying the data set to be returned.}

\item{data_items}{Names of data items to be included in the returned table.}

\item{limit}{Specify the maximum number of records to be returned, default NULL
value returns all records.}

\item{default_ord}{When selecting columns with the \code{data_items} arg, \code{FALSE}
returns the columns in the order specified.}
}
\value{
A data.frame containing the selected columns from the specified data
set.
}
\description{
Get data from a resource in tabular format.
}
\examples{
\dontrun{
get_data(
  resource = "42f17a3c-a4db-4965-ba68-3dffe6bca13a",
  data_items = c("Dose", "CumulativeNumberVaccinated"),
  limit = 10
  )
}
}
