% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_metadata.R
\name{package_metadata}
\alias{package_metadata}
\title{Get a full representation of a packages metadata.}
\usage{
package_metadata(package)
}
\arguments{
\item{package}{a valid package name or package id.}
}
\value{
a list of the package metadata.
}
\description{
Get a full representation of a specified packages metadata
as a list.
}
\examples{
\dontrun{
package_metadata(package = "gp-practice-populations")
package_metadata(package = "e3300e98-cdd2-4f4e-a24e-06ee14fcc66c")
}
}
