% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_queries.R
\name{prep_nosql_query}
\alias{prep_nosql_query}
\title{Prepare an API query without SQL.}
\usage{
prep_nosql_query(resource, fields, limit, offset)
}
\arguments{
\item{resource}{A character string specifying resource id of the data set
to be returned.}

\item{fields}{A character vector specifying the names of fields to be
included in the returned data.}

\item{limit}{A numeric value specifying the maximum number of rows to be
returned.}

\item{offset}{A numeric value specifying the number of rows to skip.}
}
\value{
A character string containing the prepared query.
}
\description{
Prepare an API query without SQL.
}
