% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_queries.R
\name{prep_sql_query}
\alias{prep_sql_query}
\title{Prepare an API query with SQL.}
\usage{
prep_sql_query(resource, fields, limit, where)
}
\arguments{
\item{resource}{A resource id identifying the data set to be returned.}

\item{fields}{Names of fields to be included in the returned table.}

\item{limit}{Specify the maximum number of records to be returned.}

\item{where}{A string containing the 'WHERE' element of a simple SQL SELECT
style query. Field names must be double quoted, non numeric values must be
single quoted, and single and double quotes must be delimited.}
}
\value{
A string containing the prepared query.
}
\description{
Prepare an API query with SQL.
}
