% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_metadata.R
\name{resource_metadata}
\alias{resource_metadata}
\title{Get metadata for a resource.}
\usage{
resource_metadata(resource)
}
\arguments{
\item{resource}{A character vector of length 1 specifying a resource id
which identifies the resource for which metadata should be returned.}
}
\value{
a list containing the resource metadata.
}
\description{
Get a specified resources metadata as a list.
}
\examples{
\dontrun{
resource_metadata(resource = "edee9731-daf7-4e0d-b525-e4c1469b8f69")
}

}
