% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_metadata.R
\name{resource_metadata}
\alias{resource_metadata}
\title{Get a table of available fields and their types for a specified resource.}
\usage{
resource_metadata(resource)
}
\arguments{
\item{resource}{a valid resource id.}
}
\value{
a data.frame detailing the names and types of all fields available
for the chosen resource.
}
\description{
Get a table of available fields and their types for a specified resource.
}
\examples{
\dontrun{
resource_metadata(resource="8f7b64b1-eb53-43e9-b888-45af0bc25505")
}

}
