% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_metadata.R
\name{resource_metadata}
\alias{resource_metadata}
\title{Get a table of available fields and their types for a specified resource.}
\usage{
resource_metadata(resource)
}
\arguments{
\item{resource}{A character string containing the resource id of the data set
to be returned.}
}
\value{
A data.frame detailing the names and types of all fields available
for the chosen resource.
}
\description{
Get a table of available fields and their types for a specified resource.
}
\examples{
\dontrun{
resource_metadata(resource="edee9731-daf7-4e0d-b525-e4c1469b8f69")
}

}
